FECore: added all FECore classes to the FECore namespace. Removed the MatrixFactory class.
git-svn-id: https://gforge.sci.utah.edu/svn/MRLProjects/FEBio@1668 b505bdfe-9d5c-0410-8600-697d5fc8e487

+++ FEBio/FESurfaceData.cpp
@@ -19,7 +19,7 @@ bool FETestData ::Save(FEDomain& dom, vector<float>& a)
 	if (dynamic_cast<FERigidSolidDomain*>(&dom) == 0)
 	{
 		int N = dom.Nodes();
-		for (int i=0; i<N; ++i) a.push_back(i);
+		for (int i=0; i<N; ++i) a.push_back((float) i);
 		return true;
 	}
 	return false;
